refactor: split prover into a generic engine + provekit-noir frontend
provekit-prover is now a field- and Noir-agnostic WHIR proving engine: it keeps
only whir_r1cs.rs (commit / prove_noir / prove_mavros / sumcheck / blinding),
operating on R1CS + a solved witness + a WhirR1CSScheme, and no longer depends
on provekit-field-bn254, acir, or nargo (cargo tree confirms).

The new provekit-noir crate owns the Noir/mavros orchestration: the Prove trait
and its impls, the witness solving (solve_witness_vec + the per-builder solve()
match), CompressedR1CS/Layers, input_utils, and logging. It depends on the
prover engine + provekit-field-bn254 and registers the field backend at the
prove entry points.

All tooling (cli/ffi/bench/wasm + passport playground) now uses provekit-noir
for proving; the prover engine stays a pure dependency. bn254 byte-identical:
field KATs, the 16 MSM witness-solving tests, and end-to-end Noir prove/verify
all pass.

Residual: the engine still links mavros for prove_mavros (frontend axis, not
field) — extracting it is a later frontend-isolation follow-up.
