+++ tooling/cli/Cargo.toml
@@ -13,7 +13,7 @@ repository.workspace = true
 provekit-common.workspace = true
 provekit-field-bn254.workspace = true
 provekit-gnark.workspace = true
-provekit-prover = { workspace = true, features = ["witness-generation", "parallel"] }
+provekit-noir = { workspace = true, features = ["witness-generation", "parallel"] }
 provekit-r1cs-compiler.workspace = true
 provekit-verifier.workspace = true
