+++ tooling/provekit-bench/Cargo.toml
@@ -12,7 +12,7 @@ repository.workspace = true
 # Workspace crates
 provekit-common.workspace = true
 provekit-field-bn254.workspace = true
-provekit-prover = { workspace = true, features = ["witness-generation", "parallel"] }
+provekit-noir = { workspace = true, features = ["witness-generation", "parallel"] }
 provekit-r1cs-compiler.workspace = true
 provekit-verifier.workspace = true
