+++ tooling/provekit-bench/tests/msm_witness_solving.rs
@@ -17,7 +17,7 @@ use {
         witness::{ConstantOrR1CSWitness, LayerScheduler, WitnessBuilder},
         FieldElement, NoirElement, TranscriptSponge,
     },
-    provekit_prover::{ec_scalar_mul, solve_witness_vec},
+    provekit_noir::{ec_scalar_mul, solve_witness_vec},
     provekit_r1cs_compiler::{
         msm::{
             add_msm_with_curve,
